+++ it/server/src/test/java/com/walmartlabs/concord/it/server/AnsibleEventIT.java
@@ -20,8 +20,6 @@ package com.walmartlabs.concord.it.server;
  * =====
  */
 
-
-
 import com.walmartlabs.concord.client.*;
 import org.junit.Test;