+++ server/plugins/ansible/impl/src/main/java/com/walmartlabs/concord/server/plugins/ansible/EventMarkerDao.java
@@ -22,20 +22,19 @@ package com.walmartlabs.concord.server.plugins.ansible;
 
 import com.walmartlabs.concord.db.AbstractDao;
 import com.walmartlabs.concord.db.MainDB;
-import com.walmartlabs.concord.db.PgUtils;
-import com.walmartlabs.concord.server.jooq.tables.EventProcessorMarkers;
-import com.walmartlabs.concord.server.jooq.tables.ProcessEventStats;
+import com.walmartlabs.concord.server.jooq.tables.EventProcessorMarker;
 import org.immutables.value.Value;
-import org.jooq.*;
+import org.jooq.Configuration;
+import org.jooq.DSLContext;
+import org.jooq.Record1;
 
 import javax.inject.Inject;
 import javax.inject.Named;
-import java.sql.Timestamp;
-import java.util.List;
 
-import static com.walmartlabs.concord.server.jooq.Tables.EVENT_PROCESSOR_MARKERS;
-import static com.walmartlabs.concord.server.jooq.Tables.PROCESS_EVENT_STATS;
-import static org.jooq.impl.DSL.*;
+import static com.walmartlabs.concord.server.jooq.Tables.EVENT_PROCESSOR_MARKER;
+import static com.walmartlabs.concord.server.jooq.Tables.PROCESS_EVENTS;
+import static org.jooq.impl.DSL.max;
+import static org.jooq.impl.DSL.value;
 
 // TODO: move to plugins sdk?
 @Named
@@ -56,60 +55,37 @@ public class EventMarkerDao extends AbstractDao {
         return super.txResult(t);
     }
 
-    public List<EventMarker> get(String processorName) {
-        ProcessEventStats e = PROCESS_EVENT_STATS.as("e");
-        EventProcessorMarkers m = EVENT_PROCESSOR_MARKERS.as("m");
+    public EventMarker get(String processorName) {
+        EventProcessorMarker m = EVENT_PROCESSOR_MARKER.as("m");
 
-        Field<Timestamp> instanceCreatedEnd = e.INSTANCE_CREATED_DATE.plus(PgUtils.interval("1 day"));
-        Field<Long> markerEventSeq = coalesce(m.EVENT_SEQ, -1L);
-        Field<Long> eventSeq = least(e.MAX_EVENT_SEQ, markerEventSeq);
+        Long currentEventSeq = txResult(tx -> tx.select(m.EVENT_SEQ)
+                .from(m)
+                .where(m.PROCESSOR_NAME.eq(processorName))
+                .fetchOne(m.EVENT_SEQ));
 
-        return txResult(tx -> tx.select(e.INSTANCE_CREATED_DATE, instanceCreatedEnd, e.MAX_EVENT_SEQ, eventSeq)
-                .from(e)
-                .leftJoin(m).on(m.PROCESSOR_NAME.eq(processorName)
-                        .and(m.INSTANCE_CREATED_DATE.eq(e.INSTANCE_CREATED_DATE)))
-                .where(e.MAX_EVENT_SEQ.greaterThan(markerEventSeq))
-                .fetch(EventMarkerDao::toMarker));
+        Long maxEventSeq = txResult(tx -> tx.select(max(PROCESS_EVENTS.EVENT_SEQ))
+                .from(PROCESS_EVENTS)
+                .fetchOne(Record1::value1));
+
+        return EventMarker.builder()
+                .eventSeq(currentEventSeq != null ? currentEventSeq : -1)
+                .maxEventSeq(maxEventSeq != null ? maxEventSeq : -1)
+                .build();
     }
 
-    public void update(DSLContext tx, String processorName, Timestamp instanceCreatedStart, long eventSeq) {
-        EventProcessorMarkers m = EVENT_PROCESSOR_MARKERS.as("m");
+    public void update(DSLContext tx, String processorName, long eventSeq) {
+        EventProcessorMarker m = EVENT_PROCESSOR_MARKER.as("m");
         tx.insertInto(m)
-                .columns(m.PROCESSOR_NAME, m.EVENT_SEQ, m.INSTANCE_CREATED_DATE)
-                .values(value(processorName), value(eventSeq), value(instanceCreatedStart))
+                .columns(m.PROCESSOR_NAME, m.EVENT_SEQ)
+                .values(value(processorName), value(eventSeq))
                 .onDuplicateKeyUpdate()
                 .set(m.EVENT_SEQ, eventSeq)
-                .where(m.PROCESSOR_NAME.eq(processorName)
-                        .and(m.INSTANCE_CREATED_DATE.eq(instanceCreatedStart)))
+                .where(m.PROCESSOR_NAME.eq(processorName))
                 .execute();
-    }
-
-    public void cleanup(String processorName) {
-        EventProcessorMarkers m = EVENT_PROCESSOR_MARKERS.as("m");
-
-        tx(tx -> tx.deleteFrom(m)
-                .where(m.PROCESSOR_NAME.eq(processorName)
-                        .and(m.INSTANCE_CREATED_DATE.notIn(
-                                tx.select(PROCESS_EVENT_STATS.INSTANCE_CREATED_DATE)
-                                        .from(PROCESS_EVENT_STATS))))
-                .execute());
-    }
-
-    private static EventMarker toMarker(Record4<Timestamp, Timestamp, Long, Long> r) {
-        return EventMarker.builder()
-                .instanceCreatedStart(r.value1())
-                .instanceCreatedEnd(r.value2())
-                .maxEventSeq(r.value3())
-                .eventSeq(r.value4())
-                .build();
     }
 
     @Value.Immutable
     public interface EventMarker {
-
-        Timestamp instanceCreatedStart();
-
-        Timestamp instanceCreatedEnd();
 
         long eventSeq();